app/requirements.txt
wheel==0.40.0
setuptools==75.1.0
sentence-transformers==2.2.2
openai==0.27.8
langchain==0.0.195
fastapi==0.96.0
uvicorn==0.22.0
sqlmodel==0.0.8
pymilvus==2.2.8
python-dotenv==1.0.0
tiktoken==0.4.0
gpt4all==0.3.0
pdfplumber
python-docx
